// Frontend/src/models/PreviousRequest.tsx
export interface PreviousRequest {
  applicants_names: string;
  _id: string;
  leader_name: string;
  requester: string;
  email: string;
  contact_no: string;
  project_title: string;
  project_description: string;
  //   Project_expenses: { type: Buffer; required: true };
  goals: string;
  risks: string;
  project_type: string;
  starting_date: string;
  ending_date: string;
  agreement: string;
  lecturer_name: string;
  lecturer_email: string;
  lecturer_response: string;
  hod_response: string;
  aces_response: string;
  reason: string;
}